feat(product): rename product-node to magento-product
