+++ libs/category/src/drivers/magento/models/get-products-response.ts
@@ -1,9 +1,9 @@
-import { ProductNode } from '@daffodil/product';
+import { MagentoProduct } from '@daffodil/product';
 import { MagentoPageInfo } from './page-info';
 
 export interface MagentoGetProductsResponse {
 	products: {
-		items: ProductNode[];
+		items: MagentoProduct[];
 		page_info: MagentoPageInfo;
 		total_count: number;
 	}
