+++ libs/category/src/drivers/magento/models/outputs/category-products-node.ts
@@ -1,0 +1,8 @@
+import { MagentoProduct } from '@daffodil/product';
+import { CategoryPageInfoNode } from './category-page-info-node';
+
+export interface CategoryProductsNode {
+  total_count: number;
+  page_info: CategoryPageInfoNode;
+  items: MagentoProduct[];
+}
